Split view for each role. Add security for Admin questions page. Modify controllers
