+++ platform_orchestrator/src/main/resources/templates/fragments/admin-navbar.html
@@ -21,7 +21,7 @@
             <a class="nav-link" href="/admin/joinRequests">Запросы на регистрацию</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link" href="/questions">Вопросы к ревью</a>
+            <a class="nav-link" href="/questionsAdmin">Вопросы к ревью</a>
         </li>
         <li class="nav-item">
             <button class="button__logout" onclick="logout()">Выйти</button>
