+++ platform_orchestrator/src/main/resources/templates/fragments/common.html
@@ -14,7 +14,7 @@
             <a class="nav-link" href="/report">Написать отчет</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link" href="#">Link</a>
+            <a class="nav-link" href="/questions">Вопросы к ревью</a>
         </li>
         <li class="nav-item">
             <a class="nav-link" href="#">Link</a>
